#11 Added ImageResource class, save scaled versions of images, use animation stack and fps counter in BoxTest2.
--HG--
extra : convert_revision : svn%3A5bd50d1f-0cf7-417a-b3c6-48be4479ebb3%40103

+++ trunk/presentation/src/se/kth/livetech/presentation/layout/Rect.java
@@ -4,7 +4,7 @@ import java.awt.Dimension;
 import java.awt.geom.Rectangle2D;
 
 public class Rect {
-	public static void setRow(Rectangle2D rect, int i, int n, Rectangle2D rowOut) {
+	public static void setRow(Rectangle2D rect, double i, int n, Rectangle2D rowOut) {
 		rowOut.setRect(
 				rect.getX(), rect.getY() + i * rect.getHeight() / n,
 				rect.getWidth(), rect.getHeight() / n);
